Added Menu for TabBar
